refactor: split recurringEventHelpers into focused modules (PR1) (#4182)
* refactor: modularize recurringEventHelpers utilities

* docs: clarify JSDoc, add interval validation, and fix byDay validation

* docs: regenerate TypeDoc documentation for modular structure

* fix: add interval validation and count=0 test case

* fix: add interval validation and achieve near 100% test coverage

* test: achieve 99% coverage with comprehensive edge case tests

* test: achieve perfect 100% statement coverage on all files

* fix: reduce timing threshold in flaky performanceTracker test

* fix: reduce all timing thresholds in performanceTracker tests

* add input validation to normalizeRecurrenceRule
