+++ docs/docs/auto-docs/utilities/recurringEvent/calculation/functions/estimateInstanceCount.md
@@ -6,7 +6,7 @@
 
 > **estimateInstanceCount**(`rule`, `estimationWindowMonths`): `number`
 
-Defined in: [src/utilities/recurringEventHelpers.ts:217](https://github.com/PalisadoesFoundation/talawa-api/tree/mainsrc/utilities/recurringEventHelpers.ts#L217)
+Defined in: [src/utilities/recurringEvent/calculation.ts:62](https://github.com/PalisadoesFoundation/talawa-api/tree/mainsrc/utilities/recurringEvent/calculation.ts#L62)
 
 Estimates the total number of instances that will be generated by a recurrence rule.
 If the rule is based on a `count`, it returns the count. If it's based on an `endDate`,
