+++ docs/docs/auto-docs/utilities/recurringEvent/formatting/functions/buildRRuleString.md
@@ -6,7 +6,7 @@
 
 > **buildRRuleString**(`recurrence`, `_startDate`): `string`
 
-Defined in: [src/utilities/recurringEventHelpers.ts:14](https://github.com/PalisadoesFoundation/talawa-api/tree/mainsrc/utilities/recurringEventHelpers.ts#L14)
+Defined in: [src/utilities/recurringEvent/formatting.ts:13](https://github.com/PalisadoesFoundation/talawa-api/tree/mainsrc/utilities/recurringEvent/formatting.ts#L13)
 
 Converts a recurrence input object into an RRULE string compliant with RFC 5545.
 This function constructs a recurrence rule string based on the provided frequency,
@@ -53,6 +53,8 @@ The recurrence input object, conforming to the recurrenceInputSchema.
 ### \_startDate
 
 `Date`
+
+Reserved for future use; currently unused but maintained for API compatibility.
 
 ## Returns
