+++ test/utilities/metrics/performanceTracker.test.ts
@@ -89,8 +89,8 @@ describe("Performance Tracker", () => {
 		const testOp = snapshot.ops["test-op"];
 		expect(testOp).toBeDefined();
 		expect(testOp?.count).toBe(1);
-		expect(testOp?.ms).toBeGreaterThanOrEqual(10);
-		expect(testOp?.max).toBeGreaterThanOrEqual(10);
+		expect(testOp?.ms).toBeGreaterThanOrEqual(5);
+		expect(testOp?.max).toBeGreaterThanOrEqual(5);
 	});
 
 	it("should handle async operation errors and still track time", async () => {
@@ -129,7 +129,7 @@ describe("Performance Tracker", () => {
 		const queryOp = snapshot.ops.query;
 		expect(queryOp).toBeDefined();
 		expect(queryOp?.count).toBe(2);
-		expect(queryOp?.ms).toBeGreaterThanOrEqual(12); // Reduced from 15ms to account for CI timing variance
+		expect(queryOp?.ms).toBeGreaterThanOrEqual(10); // Reduced from 15ms to account for CI timing variance
 	});
 
 	it("should use manual start/stop timing", async () => {
@@ -147,7 +147,7 @@ describe("Performance Tracker", () => {
 		const manualOp = snapshot.ops["manual-op"];
 		expect(manualOp).toBeDefined();
 		expect(manualOp?.count).toBe(1);
-		expect(manualOp?.ms).toBeGreaterThanOrEqual(10);
+		expect(manualOp?.ms).toBeGreaterThanOrEqual(5);
 	});
 
 	it("should track max duration for operations", async () => {
